F2 ciclo 1: ingestão de PDF unificada em caissa.ingest.pdf, do arquivo ao Document IR
Nove módulos (5.754 linhas), 166 testes, ruff e mypy limpos no pacote. Absorve o
pdf_io/pdf_text do tronco, o pdf_service do Editor e o extract do PDFimport com as
medições que os justificavam. Três portões com prova de vitalidade: primeira página de
500 em 10–21 ms; +0,4 MB em 200 páginas (a sabotagem cresce 1.198 MB); ordem de leitura
40/40 em duas colunas (a sabotagem dá 0/6). Acervo inteiro: 552 páginas, 0 falhas,
911 diagramas lidos com a via raster + F4 na GPU. Leitura lado a lado em cinco páginas
de quatro livros, fixada em test_corpus.py.

Cinco portões cegos de outras frentes, achados ao ligar a cadeia inteira, corrigidos com
teste que reprova sem a correção: a fórmula de coordenadas do Editor soma uma origem que
não existe (40 pt de erro em página girada com CropBox); o detector vetorial perdia
tabuleiros lado a lado desalinhados; SkakNew-Diagram era figurino no catálogo (Polgar
invisível); o nível 0 julgava glifos de tabuleiro como prosa (Polgar rejeitado 11/11);
todo inteiro na margem era fólio. E plain_text apagava PieceGlyph.

+++ src/caissa/ingest/pdf/geometry.py
@@ -288,8 +288,8 @@ class PageFrame:
         """
         s = self.scale_for(dpi)
         return (
-            int(math.ceil(self.width * s - _CEIL_SLACK)),
-            int(math.ceil(self.height * s - _CEIL_SLACK)),
+            math.ceil(self.width * s - _CEIL_SLACK),
+            math.ceil(self.height * s - _CEIL_SLACK),
         )
 
     def page_to_pixels(
